agent_backstage/backstage_mcp/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "backstage"
version = "0.1.0"
description = "Generated MCP server for backstage"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "*.py" },
]

[tool.poetry.dependencies]
python = ">=3.8.1,<4.0"
httpx = "^0.24.0"
python-dotenv = "^1.0.0"
pydantic = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^23.0"
isort = "^5.0"
flake8 = "^6.0"
